On translated content, invalid call to EntityTypeManager::getTranslationFromContext
On translated content, call to EntityTypeManager that is not injected in GraphQLEntityRow


I've translated content (no filter), I set up a very simple views:

```query test {
	graphqlViewsTestGraphql1View {
    results {
      entityLabel
    }
	}
}```

I got this error:

```
Error: Call to undefined method Drupal\Core\Entity\EntityTypeManager::getTranslationFromContext() in Drupal\graphql_views\Plugin\views\row\GraphQLEntityRow->getEntityTranslation() (line 77 of /var/www/project/web/core/modules/views/src/Entity/Render/EntityTranslationRenderTrait.php) 
#0 /var/www/project/web/modules/contrib/graphql_views/src/Plugin/views/row/GraphQLEntityRow.php(89):
```

GraphQLEntityRow use the trait EntityTranslationRenderTrait that make use of `$this->getEntityManager` that is supposed to be an EntityManager. But this is a EntityTypeManager that is passed and this class does not implement getTranslationFromContext. 

Furthermore, EntityManager::getTranslationFromContext is deprecated in favor of EntityRepository. I think this is the reason behind the use of EntityTypeManager (used for entityManager->getDefinition in EntityTranslationRenderTrait L52). 

But as we depends on EntityTranslationRenderTrait that use $this->getEntityManager for multiple cases (EntityTypeManager / EntityRepository), I think we can't get rid of the use of the deprecated EntityManager.